Removes conflicting password_digest column from users
When upgrading to Rails 5 a bug surfaced that ended up being tied to a
change in Devise 4.2 which now uses #serializable_hash on #inspect to
avoid serializing sensitive data.  This new code calls
\#password_digest(password) as an attribute when it is a method, causing
an "ArgumentError (given 0, expected 1)" error.

password_digest is a vestige of the original code that used ActiveRecord
validations for password.  It is no longer needed with Devise and since
it is now causing this error, this commit removes it.