Fixes sign out bug by changing the method to :get
This looks like it might be a subtle change introduced by Rails 5. The
sign out operation used to be a call to destroy_user_session_path,
method: :delete but now the method is :get

The sign out test was failing and I know it was working prior to the
upgrade.