Fixes problem with visit_counter test
The Ahoy package was not registering visits when under test mode.  It
turned out that there is code in the package that checks for "bots" and
that code interpreted test mode as being a bot visit.  In addition to
the final test code, this commit also patches the Ahoy
ActiveRecordTokenStore#exclude? method to allow test mode visits to be
stored.