Fixes identity swap bug and adds update flash msgs
Problem: When an admin edited the attributes of another user, the
account became signed in as that user when the Update button was hit.

Root Cause: was doing a `bypass_sign_in` for the current user so that a
password change would not log the user out.  But, when editing a user
that is not the current user, the `bypass_sign_in` call logged in the
target user.  Adding a conditional to prevent this from being called in
the "other user" scenario fixed the problem.

This commit fixes that problem and adds flash messages for successful
and failed update attempts.  Also, tests.