Improves seed data, specifically:
* Add New Years and NYE as named holidays
* Makes NY and NYE to have Markdown links to show name regularizing

When the clone/copy gig feature was added there was no way to tell at a
glance if the uniq-ifying code was combining like holiday names.  By
making these start/end-of-year holidays explicit, there is now a test
case.

Subsequently, it was discovered that many gig names are [MD](links.com)
and while the name may match, the URL would differ, so code was added to
regularize these so they would telescope into matching holidays and
NY/NYE were made to be links since we know that every year has one of
each and they are the same (except for the random link).